requirements.txt
Flask
gunicorn
pyzbar
PyPDF2
PyMuPDF
Pillow
requests
selenium
webdriver-manager
beautifulsoup4
flask-cors
pytesseract
